[INS-06] Refactored changes for new directory structure
