+++ inetsense/inetsense-server/inetsense-server-collector/src/main/java/hu/elte/inetsense/server/collector/util/JsonValidationException.java
@@ -1,0 +1,11 @@
+package hu.elte.inetsense.server.collector.util;
+
+/**
+ * Created by balintkiss on 5/1/16.
+ */
+public class JsonValidationException extends Exception {
+
+    public JsonValidationException(String message) {
+        super(message);
+    }
+}
